feat: display correct format date in the input
