--- src/components/customer-data/custom-date-picker/CustomDayPicker.css
@@ -1,8 +1,0 @@
-.DayPickerInput-OverlayWrapper {
-  right: 90px;
-}
-
-.react-time-picker__inputGroup__input.react-time-picker__inputGroup__hour,
-.react-time-picker__inputGroup__input.react-time-picker__inputGroup__minute {
-  border: none;
-}
