--- src/components/customer-data/custom-date-picker/index.js
@@ -1,3 +1,0 @@
-import CustomDayPicker from './CustomDayPicker'
-
-export default CustomDayPicker
