+++ src/components/delivery-date-input/index.js
@@ -1,0 +1,3 @@
+import DeliveryDateInput from './DeliveryDateInput'
+
+export default DeliveryDateInput
